[TASK] Update codeception to latest patch level
composer update codeception/codeception --with-dependencies

Updates codeception from 2.4.1 to 2.4.4 and brings
phpunit 7.1.4 to 7.1.5 along with some deeper dependencies.
The younger phpunit bundle seems to be a bit more picky on
string assertions, so we need to fix two unit tests that
silently used the bug before.

Resolves: #85623
Releases: master
Reviewed-on: https://review.typo3.org/57662
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/core/Tests/Unit/Resource/Driver/LocalDriverTest.php
@@ -643,7 +643,7 @@ class LocalDriverTest extends BaseTestCase
         list($basedir, $subject) = $this->prepareRealTestEnvironment();
         mkdir($basedir . '/someDir');
         $subject->createFile('testfile.txt', '/someDir');
-        $this->assertEquals($testpattern, decoct(fileperms($basedir . '/someDir/testfile.txt') & 0777));
+        $this->assertEquals((int)$testpattern, (int)(decoct(fileperms($basedir . '/someDir/testfile.txt') & 0777)));
     }
 
     /**********************************
